refactor(article): correct package name to "nettee.blolet.article.web"
