+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/ArticleQueryApi.java
@@ -1,11 +1,11 @@
-package nettee.blole.article.web;
+package nettee.blolet.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import nettee.blole.article.web.dto.ArticleQueryDto.ArticleDetailViewResponse;
-import nettee.blole.article.web.dto.ArticleQueryDto.ArticleListViewsResponse;
 import nettee.blolet.article.application.usecase.ArticleReadUseCase;
+import nettee.blolet.article.web.dto.ArticleQueryDto.ArticleDetailViewResponse;
+import nettee.blolet.article.web.dto.ArticleQueryDto.ArticleListViewsResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
