+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/DraftBlockQueryApi.java
@@ -1,16 +1,16 @@
-package nettee.blole.article.web;
+package nettee.blolet.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import nettee.blole.article.web.dto.DraftBlockQueryDto.DraftBlockDetailResponse;
 import nettee.blolet.article.application.usecase.DraftBlockReadByStatusUseCase;
 import nettee.blolet.article.application.usecase.DraftBlockReadUseCase;
 import nettee.blolet.article.domain.sub.DraftBlockStatus;
 import nettee.blolet.article.readmodel.DraftBlockReadModels.DraftBlockDetail;
 import nettee.blolet.article.readmodel.DraftBlockReadModels.DraftBlockSummary;
+import nettee.blolet.article.web.dto.DraftBlockQueryDto.DraftBlockDetailResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
