+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/DraftQueryApi.java
@@ -1,18 +1,18 @@
-package nettee.blole.article.web;
+package nettee.blolet.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import nettee.blole.article.web.dto.DraftQueryDto.DraftDetailResponse;
-import nettee.blole.article.web.dto.DraftQueryDto.DraftTitleResponse;
 import nettee.blolet.article.application.usecase.DraftReadByStatusesUseCase;
 import nettee.blolet.article.application.usecase.DraftReadUseCase;
 import nettee.blolet.article.domain.sub.DraftStatus;
 import nettee.blolet.article.readmodel.DraftReadModels.DraftDetail;
 import nettee.blolet.article.readmodel.DraftReadModels.DraftSummary;
 import nettee.blolet.article.readmodel.DraftReadModels.DraftTitle;
+import nettee.blolet.article.web.dto.DraftQueryDto.DraftDetailResponse;
+import nettee.blolet.article.web.dto.DraftQueryDto.DraftTitleResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
