+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/dto/DraftCommandDto.java
@@ -1,4 +1,4 @@
-package nettee.blole.article.web.dto;
+package nettee.blolet.article.web.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
