+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/dto/DraftQueryDto.java
@@ -1,4 +1,4 @@
-package nettee.blole.article.web.dto;
+package nettee.blolet.article.web.dto;
 
 import lombok.Builder;
 import nettee.blolet.article.readmodel.DraftReadModels.DraftDetail;
