+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/dto/SeriesQueryDto.java
@@ -1,4 +1,4 @@
-package nettee.blole.article.web.dto;
+package nettee.blolet.article.web.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
