+++ services/article/driving/web-mvc/src/main/java/nettee/blolet/article/web/mapper/DraftBlockDtoMapper.java
@@ -1,9 +1,9 @@
-package nettee.blole.article.web.mapper;
+package nettee.blolet.article.web.mapper;
 
-import nettee.blole.article.web.dto.DraftBlockCommandDto.DraftBlockCreateCommand;
-import nettee.blole.article.web.dto.DraftBlockCommandDto.DraftBlockUpdateCommand;
 import nettee.blolet.article.domain.DraftBlock;
 import nettee.blolet.article.domain.sub.DraftBlockStatus;
+import nettee.blolet.article.web.dto.DraftBlockCommandDto.DraftBlockCreateCommand;
+import nettee.blolet.article.web.dto.DraftBlockCommandDto.DraftBlockUpdateCommand;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
